admin-panel/vizual_proqramlasdirma_version2/dao-b1-main/database/requirements.txt
# Database dependencies for PostgreSQL
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
alembic>=1.13.0
python-dotenv>=1.0.0
